Fix some wxMotif build warnings about deprecated methods.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@18811 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/motif/choice.h
@@ -94,7 +94,7 @@ protected:
     // common part of all contructors
     void Init();
 
-    int           m_noStrings;
+    size_t        m_noStrings;
     WXWidget      m_menuWidget;
     WXWidget      m_buttonWidget;
     wxWidgetArray m_widgetArray;
